Add Register on behalf tab to Manage page
Deletes standalone organizer_checkin.rs page and lifts its player/team
registration forms into a new "Register on behalf" tab on the Manage page.
The existing Registrations tab is preserved unchanged. Bootstrap data
(teams, waiver_required, waiver_url) comes from TournamentManageResponse
already fetched by Manage. Successful submissions trigger refresh so the
Registrations tab stays in sync.
