+++ frontend/src/main.rs
@@ -89,9 +89,6 @@ enum Route {
     #[route("/:url/register")]
     TournamentRegister { url: String },
 
-    #[route("/:url/checkin")]
-    OrganizerCheckin { url: String },
-
     #[route("/:url/manage")]
     Manage { url: String },
