+++ frontend/src/pages/layout.rs
@@ -62,7 +62,6 @@ fn page_title_for_route(route: &Route) -> String {
         Route::License { .. } => "License".into(),
         Route::ArctosScheduleScript { .. } => "Arctos Schedule Script".into(),
         Route::DataAccessibilityGuide { .. } => "Data Accessibility Guide".into(),
-        Route::OrganizerCheckin { url } => format!("{url} Event Check-in"),
     }
 }
