+++ frontend/src/pages/mod.rs
@@ -3,7 +3,6 @@ mod add_injury;
 mod arctos_schedule_script;
 mod bracket;
 mod data_accessibility_guide;
-mod organizer_checkin;
 mod docs;
 mod edit_injury;
 mod edit_player_profile;
@@ -55,7 +54,6 @@ pub use add_injury::AddInjury;
 pub use arctos_schedule_script::ArctosScheduleScript;
 pub use bracket::Bracket;
 pub use data_accessibility_guide::DataAccessibilityGuide;
-pub use organizer_checkin::OrganizerCheckin;
 pub use docs::Docs;
 pub use edit_injury::EditInjury;
 pub use edit_player_profile::EditPlayerProfile;
